added all fields, not staying because not in DB yet
