+++ src/components/IndividualPlant.js
@@ -38,6 +38,10 @@ export const IndividualPlant = ({plantData, openImageModal})=>{
  
                                 <ListGroupItem><span style={{fontWeight:'bold'}}>Hardiness:</span> {plantData.hardiness}</ListGroupItem>
                                 <ListGroupItem><span style={{fontWeight:'bold'}}>Water:</span> {plantData.water}</ListGroupItem>
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Watering Time:</span> {plantData.waterTime}</ListGroupItem>
+
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Watering Days:</span> {plantData.waterDay}</ListGroupItem>
+
  
  
                             </ListGroup>
